# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Lumière Beauty API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: UserRole;
  skinType?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  accessToken: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  skinType?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  category: string;
  subcategory: string;
  description: string;
  price: number;
  salePrice?: number | null;
  images: string[];
  rating: number;
  reviewCount: number;
  stock: number;
  skinTypes: string[];
  concerns: string[];
  isBestSeller: boolean;
  isNewArrival: boolean;
  isFeatured: boolean;
  ingredients?: string | null;
  howToUse?: string | null;
}

export type ReviewSource = (typeof ReviewSource)[keyof typeof ReviewSource];

export const ReviewSource = {
  user: "user",
  scraped: "scraped",
} as const;

export interface Review {
  id: number;
  productId: number;
  userId?: number | null;
  reviewerName: string;
  rating: number;
  title?: string | null;
  body: string;
  skinType?: string | null;
  source: ReviewSource;
  isApproved: boolean;
  createdAt: string;
}

export type ProductDetail = Product & {
  reviews?: Review[];
};

export interface ProductsResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Category {
  name: string;
  subcategories: string[];
}

export interface CreateProductBody {
  name: string;
  slug: string;
  category: string;
  subcategory: string;
  description: string;
  ingredients?: string | null;
  howToUse?: string | null;
  price: number;
  salePrice?: number | null;
  images: string[];
  stock: number;
  skinTypes: string[];
  concerns: string[];
  isBestSeller: boolean;
  isNewArrival: boolean;
  isFeatured: boolean;
}

export interface GeneratedDescriptionResponse {
  description: string;
}

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  product: Product;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  shipping: number;
  total: number;
  itemCount: number;
}

export interface CartItemBody {
  productId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  price: number;
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Address {
  id?: number;
  fullName: string;
  street: string;
  city: string;
  state: string;
  zip: string;
  country: string;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  total: number;
  shippingAddress: Address;
  createdAt: string;
  userId: number;
  userEmail?: string | null;
}

export interface CreateOrderBody {
  shippingAddress: Address;
  discountCode?: string | null;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface AddressBody {
  fullName: string;
  street: string;
  city: string;
  state: string;
  zip: string;
  country: string;
}

export interface CreateReviewBody {
  rating: number;
  title?: string | null;
  body: string;
  skinType?: string | null;
}

export interface QuizAnswers {
  skinType: string;
  concerns: string[];
  ageRange: string;
  dailyTime: string;
  budget: string;
}

export interface QuizRecommendation {
  product: Product;
  reason: string;
}

export interface QuizResult {
  skinProfile: QuizAnswers;
  recommendations: QuizRecommendation[];
  createdAt?: string | null;
}

export interface WishlistBody {
  productId: number;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatMessageBody {
  message: string;
  history?: ChatMessage[];
}

export interface ChatResponse {
  reply: string;
}

export interface UpdateProfileBody {
  firstName?: string;
  lastName?: string;
  skinType?: string | null;
}

export interface AdminStats {
  totalRevenue: number;
  ordersToday: number;
  totalProducts: number;
  totalCustomers: number;
  recentOrders: Order[];
  lowStockCount: number;
}

export type SalesAnalyticsRevenueByMonthItem = {
  month: string;
  revenue: number;
};

export type SalesAnalyticsSalesByCategoryItem = {
  category: string;
  sales: number;
};

export type SalesAnalyticsTopProductsItem = {
  name: string;
  sales: number;
};

export interface SalesAnalytics {
  revenueByMonth: SalesAnalyticsRevenueByMonthItem[];
  salesByCategory: SalesAnalyticsSalesByCategoryItem[];
  topProducts: SalesAnalyticsTopProductsItem[];
}

export type GetProductsParams = {
  category?: string;
  subcategory?: string;
  skinType?: string;
  minPrice?: number;
  maxPrice?: number;
  minRating?: number;
  inStock?: boolean;
  sort?: GetProductsSort;
  search?: string;
  page?: number;
  limit?: number;
};

export type GetProductsSort =
  (typeof GetProductsSort)[keyof typeof GetProductsSort];

export const GetProductsSort = {
  bestSelling: "bestSelling",
  newArrivals: "newArrivals",
  priceLow: "priceLow",
  priceHigh: "priceHigh",
  topRated: "topRated",
} as const;