# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lumière Beauty API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/products:
    get:
      operationId: adminGetProducts
      tags: [admin]
      summary: Admin list all products
      responses:
        "200":
          description: All products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsResponse"
    post:
      operationId: adminCreateProduct
      tags: [admin]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/products/{id}/generate-description:
    post:
      operationId: generateProductDescription
      tags: [admin]
      summary: AI-generate product description
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generated description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedDescriptionResponse"

  # CART
...
  /admin/orders/{id}/status:
    put:
      operationId: adminUpdateOrderStatus
      tags: [admin]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # REVIEWS
...
  /account/profile:
    get:
      operationId: getProfile
      tags: [account]
      summary: Get user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateProfile
      tags: [account]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /account/addresses:
    get:
      operationId: getAddresses
      tags: [account]
      summary: Get saved addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: addAddress
      tags: [account]
      summary: Add address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressBody"
      responses:
        "201":
          description: Address added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"

  # ADMIN DASHBOARD
...
  /admin/reviews/{id}/approve:
    put:
      operationId: adminApproveReview
      tags: [admin]
      summary: Approve review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/reviews/{id}:
    delete:
      operationId: adminDeleteReview
      tags: [admin]
      summary: Delete review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [customer, admin]
        skinType:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, email, firstName, lastName, role, createdAt]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        subcategory:
          type: string
        description:
          type: string
        price:
          type: number
        salePrice:
          type: number
          nullable: true
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        stock:
          type: integer
        skinTypes:
          type: array
          items:
            type: string
        concerns:
          type: array
          items:
            type: string
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
        isFeatured:
          type: boolean
        ingredients:
          type: string
          nullable: true
        howToUse:
          type: string
          nullable: true
      required: [id, name, slug, category, subcategory, description, price, images, rating, reviewCount, stock, skinTypes, concerns, isBestSeller, isNewArrival, isFeatured]
...
    ProductsResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [products, total, page, totalPages]
...
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        category:
          type: string
        subcategory:
          type: string
        description:
          type: string
        ingredients:
          type: string
          nullable: true
        howToUse:
          type: string
          nullable: true
        price:
          type: number
        salePrice:
          type: number
          nullable: true
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        skinTypes:
          type: array
          items:
            type: string
        concerns:
          type: array
          items:
            type: string
        isBestSeller:
          type: boolean
        isNewArrival:
          type: boolean
        isFeatured:
          type: boolean
      required: [name, slug, category, subcategory, description, price, images, stock, skinTypes, concerns, isBestSeller, isNewArrival, isFeatured]
    GeneratedDescriptionResponse:
      type: object
      properties:
        description:
          type: string
      required: [description]
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        price:
          type: number
        quantity:
          type: integer
      required: [id, productId, productName, productImage, price, quantity]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shipping:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
        userId:
          type: integer
        userEmail:
          type: string
          nullable: true
      required: [id, orderNumber, status, items, subtotal, shipping, total, shippingAddress, createdAt, userId]
...
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
      required: [status]
    Address:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
      required: [fullName, street, city, state, zip, country]
    AddressBody:
      type: object
      properties:
        fullName:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
      required: [fullName, street, city, state, zip, country]
...
    UpdateProfileBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        skinType:
          type: string
          nullable: true
...